1271 data broker convert mocha to jest (#1432)
* Rename tests folder to __tests__

* Add jest.config.json

* Update package.json

* Convert mocha to jest

* Delete .babelrc

* Convert mocha to jest

* Add eslint config in package.json

* Change mockResolved to mockReturn

* Remove mockClear

* Use createJestMockInstance in TupaiaService.stub.js

* Reinstate script order

* Data broker test tweaks (#1485)
