--- packages/data-broker/.babelrc
@@ -1,7 +1,0 @@
-{
-  "env": {
-    "test": {
-      "plugins": ["istanbul"]
-    }
-  }
-}
