+++ packages/data-broker/package.json
@@ -15,9 +15,9 @@
     "build:ts": "tsc",
     "lint": "eslint --ignore-path ../../.gitignore .",
     "lint:fix": "yarn lint --fix .",
-    "test": "mocha",
-    "test:coverage": "cross-env NODE_ENV=test nyc mocha",
-    "test:debug": "mocha --inspect-brk"
+    "test": "jest",
+    "test:coverage": "yarn test --coverage",
+    "test:debug": "yarn test --debug"
   },
   "dependencies": {
     "@tupaia/data-api": "1.0.0",
@@ -33,5 +33,8 @@
     "lodash.pickby": "^4.6.0",
     "moment": "^2.24.0",
     "winston": "^3.3.3"
+  },
+  "eslintConfig": {
+    "extends": "@beyondessential/jest"
   }
 }
