+++ packages/data-broker/src/__tests__/services/dhis/DhisService/DhisService.test.js
@@ -1,0 +1,17 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+import { cleanupDhisApiStub, setupDhisApiForStubbing } from './DhisService.stubs';
+import { testDelete } from './testDelete';
+import { testPull } from './testPull';
+import { testPush } from './testPush';
+
+describe('DhisService', () => {
+  describe('push()', testPush);
+
+  describe('delete()', testDelete);
+
+  describe('pull()', testPull);
+});
