+++ packages/data-broker/src/__tests__/services/tupaia/TupaiaService.stubs.js
@@ -1,0 +1,27 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+import { DATA_ELEMENTS } from './TupaiaService.fixtures';
+import { createJestMockInstance } from '../../../../../utils/src/testUtilities';
+
+export const createModelsStub = () => ({
+  dataSource: {
+    getTypes: () => ({ DATA_ELEMENT: 'dataElement', DATA_GROUP: 'dataGroup' }),
+  },
+});
+
+export const createTupaiaDataApiStub = ({
+  fetchAnalyticsResponse = [],
+  fetchEventsResponse = [],
+} = {}) =>
+  createJestMockInstance('@tupaia/data-api', 'TupaiaDataApi', {
+    fetchAnalytics: jest.fn().mockResolvedValue(fetchAnalyticsResponse),
+    fetchEvents: jest.fn().mockResolvedValue(fetchEventsResponse),
+    fetchDataElements: jest
+      .fn()
+      .mockImplementation(async dataElementCodes =>
+        dataElementCodes.map(code => DATA_ELEMENTS[code]).filter(de => de),
+      ),
+  });
