--- packages/data-broker/src/tests/services/dhis/DhisService/DhisService.test.js
@@ -1,25 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-import { cleanupDhisApiStub, setupDhisApiForStubbing } from './DhisService.stubs';
-import { testDelete } from './testDelete';
-import { testPull } from './testPull';
-import { testPush } from './testPush';
-
-describe('DhisService', () => {
-  before(() => {
-    setupDhisApiForStubbing();
-  });
-
-  after(() => {
-    cleanupDhisApiStub();
-  });
-
-  describe('push()', testPush);
-
-  describe('delete()', testDelete);
-
-  describe('pull()', testPull);
-});
